# docs/openapi.yaml
basePath: /api/v1
definitions:
...
  github_com_uigraph_app_internal_httputil.errorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
  description: REST API for uigraph.
  title: uigraph API
  version: "1.0"
paths:
...
  /orgs/{orgID}/diagrams:
    get:
      parameters:
      - description: orgID
        in: path
        name: orgID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
      security:
      - BearerAuth: []
      summary: List
      tags:
      - diagrams
    post:
      parameters:
      - description: orgID
        in: path
        name: orgID
        required: true
        type: string
      - description: request body
        in: body
        name: body
        schema:
          type: object
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
      security:
      - BearerAuth: []
      summary: Create
      tags:
      - diagrams
  /orgs/{orgID}/diagrams/{diagramID}:
    delete:
      parameters:
      - description: orgID
        in: path
        name: orgID
        required: true
        type: string
      - description: diagramID
        in: path
        name: diagramID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
      security:
      - BearerAuth: []
      summary: Delete
      tags:
      - diagrams
    get:
      parameters:
      - description: orgID
        in: path
        name: orgID
        required: true
        type: string
      - description: diagramID
        in: path
        name: diagramID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
      security:
      - BearerAuth: []
      summary: Get
      tags:
      - diagrams
    put:
      parameters:
      - description: orgID
        in: path
        name: orgID
        required: true
        type: string
      - description: diagramID
        in: path
        name: diagramID
        required: true
        type: string
      - description: request body
        in: body
        name: body
        schema:
          type: object
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
      security:
      - BearerAuth: []
      summary: Update
      tags:
      - diagrams
  /orgs/{orgID}/diagrams/{diagramID}/content:
    get:
      parameters:
      - description: orgID
        in: path
        name: orgID
        required: true
        type: string
      - description: diagramID
        in: path
        name: diagramID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
      security:
      - BearerAuth: []
      summary: GetContent
      tags:
      - diagrams
...
  /orgs/{orgID}/diagrams/{diagramID}/thumbnail:
    post:
      parameters:
      - description: orgID
        in: path
        name: orgID
        required: true
        type: string
      - description: diagramID
        in: path
        name: diagramID
        required: true
        type: string
      - description: request body
        in: body
        name: body
        schema:
          type: object
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
      security:
      - BearerAuth: []
      summary: UpdateThumbnail
      tags:
      - diagrams
  /orgs/{orgID}/diagrams/{diagramID}/thumbnail/confirm:
    post:
      parameters:
      - description: orgID
        in: path
        name: orgID
        required: true
        type: string
      - description: diagramID
        in: path
        name: diagramID
        required: true
        type: string
      - description: request body
        in: body
        name: body
        schema:
          type: object
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
      security:
      - BearerAuth: []
      summary: ConfirmThumbnailUpload
      tags:
      - diagrams
  /orgs/{orgID}/diagrams/{diagramID}/thumbnail/prepare:
    post:
      parameters:
      - description: orgID
        in: path
        name: orgID
        required: true
        type: string
      - description: diagramID
        in: path
        name: diagramID
        required: true
        type: string
      - description: request body
        in: body
        name: body
        schema:
          type: object
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
      security:
      - BearerAuth: []
      summary: PrepareThumbnailUpload
      tags:
      - diagrams
...
  /orgs/{orgID}/diagrams/sync:
    post:
      parameters:
      - description: orgID
        in: path
        name: orgID
        required: true
        type: string
      - description: request body
        in: body
        name: body
        schema:
          type: object
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
      security:
      - BearerAuth: []
      summary: Sync
      tags:
      - diagrams
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"